# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: cdCTF API
  version: 0.1.0
  description: cdCTF API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ctf:
    get:
      operationId: listCtfChallenges
      tags: [ctf]
      summary: List CTF challenges
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [easy, medium, hard, insane]
        - in: query
          name: solved
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: minPoints
          schema:
            type: integer
        - in: query
          name: maxPoints
          schema:
            type: integer
      responses:
        "200":
          description: List of CTF challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CtfChallenge"
...
  /ctf/{id}:
    get:
      operationId: getCtfChallenge
      tags: [ctf]
      summary: Get a CTF challenge by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CTF challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtfChallengeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /competitions/{id}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Get competition detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/users/{id}/block:
    post:
      operationId: adminBlockUser
      tags: [admin]
      summary: Block a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/unblock:
    post:
      operationId: adminUnblockUser
      tags: [admin]
      summary: Unblock a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/ctf:
    post:
      operationId: adminCreateCtf
      tags: [admin]
      summary: Create a CTF challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCtfBody"
      responses:
        "201":
          description: CTF created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtfChallengeDetail"
  /admin/ctf/{id}:
    put:
      operationId: adminUpdateCtf
      tags: [admin]
      summary: Update a CTF challenge
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCtfBody"
      responses:
        "200":
          description: CTF updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtfChallengeDetail"

    delete:
      operationId: adminDeleteCtf
      tags: [admin]
      summary: Delete a CTF challenge
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CTF deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/ctf/{id}/unblock-user/{userId}:
    post:
      operationId: adminUnblockCtfUser
      tags: [admin]
      summary: Unblock a user from a specific CTF
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unblocked for CTF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/blocked-tasks/{type}/{taskId}/unblock/{userId}:
    post:
      operationId: adminUnblockTask
      tags: [admin]
      summary: Unblock a task for a user
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [ctf, lesson]
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/competitions:
    post:
      operationId: adminCreateCompetition
      tags: [admin]
      summary: Create a competition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitionBody"
      responses:
        "201":
          description: Competition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /admin/competitions/{id}:
    put:
      operationId: adminUpdateCompetition
      tags: [admin]
      summary: Update a competition
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetitionBody"
      responses:
        "200":
          description: Competition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /admin/competitions/{id}/users/{userId}:
    post:
      operationId: adminAddCompetitionUser
      tags: [admin]
      summary: Add user to private competition
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/lessons:
    post:
      operationId: adminCreateLesson
      tags: [admin]
      summary: Create a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
  /admin/lessons/{id}:
    put:
      operationId: adminUpdateLesson
      tags: [admin]
      summary: Update a lesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "200":
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"

    delete:
      operationId: adminDeleteLesson
      tags: [admin]
      summary: Delete a lesson
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
          nullable: true
        points:
          type: integer
        role:
          type: string
          enum: [user, admin]
        emailVerified:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, nickname, email, points, role, emailVerified, isBlocked, createdAt]

    # CTF
    CtfChallenge:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, insane]
        points:
          type: integer
        solvedCount:
          type: integer
        isSolved:
          type: boolean
        isBlocked:
          type: boolean
      required: [id, name, category, difficulty, points, solvedCount, isSolved, isBlocked]
    CtfChallengeDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameUz:
          type: string
          nullable: true
        nameRu:
          type: string
          nullable: true
        description:
          type: string
        descriptionUz:
          type: string
          nullable: true
        descriptionRu:
          type: string
          nullable: true
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, insane]
        points:
          type: integer
        hintCost:
          type: integer
        fileUrl:
          type: string
          nullable: true
        solvedCount:
          type: integer
        isSolved:
          type: boolean
        isBlocked:
          type: boolean
        hintUsed:
          type: boolean
        wrongAttempts:
          type: integer
      required: [id, name, description, category, difficulty, points, hintCost, solvedCount, isSolved, isBlocked, hintUsed, wrongAttempts]
...
    Competition:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended]
        participantCount:
          type: integer
        ctfCount:
          type: integer
        isJoined:
          type: boolean
      required: [id, name, type, startTime, endTime, status, participantCount, ctfCount, isJoined]
    CompetitionDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, ended]
        participantCount:
          type: integer
        isJoined:
          type: boolean
        challenges:
          type: array
          items:
            $ref: "#/components/schemas/CtfChallenge"
        certificateUrl:
          type: string
          nullable: true
      required: [id, name, type, startTime, endTime, status, participantCount, isJoined, challenges]

    # Profile
...
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
...
    CreateCtfBody:
      type: object
      properties:
        name:
          type: string
        nameUz:
          type: string
          nullable: true
        nameRu:
          type: string
          nullable: true
        description:
          type: string
        descriptionUz:
          type: string
          nullable: true
        descriptionRu:
          type: string
          nullable: true
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, insane]
        points:
          type: integer
        hint:
          type: string
          nullable: true
        flag:
          type: string
        fileUrl:
          type: string
          nullable: true
      required: [name, description, category, difficulty, points, flag]
    CreateCompetitionBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [public, private]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        ctfIds:
          type: array
          items:
            type: integer
      required: [name, type, startTime, endTime, ctfIds]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        titleUz:
          type: string
          nullable: true
        titleRu:
          type: string
          nullable: true
        content:
          type: string
        contentUz:
          type: string
          nullable: true
        contentRu:
          type: string
          nullable: true
        categoryId:
          type: integer
        points:
          type: integer
        questions:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              questionUz:
                type: string
                nullable: true
              questionRu:
                type: string
                nullable: true
              options:
                type: array
                items:
                  type: string
              optionsUz:
                type: array
                items:
                  type: string
                nullable: true
              optionsRu:
                type: array
                items:
                  type: string
                nullable: true
              correctOption:
                type: integer
            required: [question, options, correctOption]
      required: [title, content, categoryId, points, questions]